UI tests refactor (#180)
Ui tests refactor
