--- src/test/ui/configs/.mocharc-debug.js
@@ -1,3 +1,0 @@
-module.exports = {
-    timeout: 99999999
-}
