--- src/test/ui/configs/settings.json
@@ -1,4 +1,0 @@
-{
-    "files.simpleDialog.enable": true,
-    "window.dialogStyle": "custom"
-}
