--- src/uiTests/models/IBreakpoint.ts
@@ -1,7 +1,0 @@
-/**
- * VsCode breakpoint.
- */
-export interface IBreakpoint {
-    fileName: string;
-    lineNumber: number;
-}
