--- src/uiTests/models/ICcppConfiguration.ts
@@ -1,12 +1,0 @@
-/**
- * C/C++ configuration template.
- */
-export interface ICcppConfiguration {
-    name: string;
-    includePath: [];
-    defines: [];
-    compilerPath: string;
-    cStandard: string;
-    cppStandard: string;
-    intelliSenseMode: string;
-}
