--- src/uiTests/models/INotification.ts
@@ -1,8 +1,0 @@
-/**
- * VsCode notification template.
- */
-export interface INotification {
-    name: string;
-    message: string;
-    installButton: string;
-}
