--- src/uiTests/models/IVsCodeTask.ts
@@ -1,9 +1,0 @@
-/**
- * VsCode task template.
- */
-export interface IVsCodeTask {
-    label: string;
-    command: string;
-    type: string;
-    problemMatcher?: [];
-}
