--- src/uiTests/models/index.ts
@@ -1,8 +1,0 @@
-import { ICcppConfiguration } from "../models/ICcppConfiguration";
-import { INotification } from "../models/INotification";
-import { IVsCodeTask } from "../models/IVsCodeTask";
-import { IThread } from "../models/IThread";
-import { IBreakpoint } from "../models/IBreakpoint";
-import { IConditionalBreakpoint } from "../models/IConditionalBreakpoint";
-
-export { ICcppConfiguration, INotification, IVsCodeTask, IThread, IBreakpoint, IConditionalBreakpoint };
