--- src/uiTests/utils/Enums.ts
@@ -1,17 +1,0 @@
-export const enum ThreadProperties {
-    Id = "Id",
-    Location = "Location"
-}
-
-export enum ConditionalBreakpointTypes {
-    SimdCommand = "SimdCommand",
-    SimdGui = "SimdGui",
-    NativeCommand = "NativeCommand",
-    NativeGui = "NativeGui",
-}
-
-export enum OneApiDebugPane {
-    OneApiGpuThreads = "OneApiGpuThreads",
-    HardwareInfo = "HardwareInfo",
-    SelectedLane = "SelectedLane"
-}
